requirements.txt
streamlit
google-generativeai
python-dotenv
requests
openai
google-generativeai
pillow
python-dotenv